+++ infra/config-store/src/main/java/com/simprints/infra/config/store/local/migrations/models/OldProjectConfig.kt
@@ -98,7 +98,7 @@ internal data class OldProjectConfig(
                 rankOne = FaceConfiguration.FaceSdkConfiguration(
                     nbOfImagesToCapture = faceNbOfFramesCaptured?.toIntOrNull()
                         ?: DEFAULT_FACE_FRAMES_TO_CAPTURE,
-                    qualityThreshold = faceQualityThreshold.toInt(),
+                    qualityThreshold = faceQualityThreshold.toFloat(),
                     imageSavingStrategy = if (saveFaceImages.toBoolean()) {
                         FaceConfiguration.ImageSavingStrategy.ONLY_USED_IN_REFERENCE
                     } else {
